requirements.txt
asgiref==3.5.2
astroid==2.11.6
autopep8==1.6.0
black==22.6.0
certifi==2022.6.15
cffi==1.15.1
chardet==5.0.0
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.5
colorclass==2.2.2
cryptography==37.0.4
defusedxml==0.7.1
dill==0.3.5.1
dj-database-url==0.5.0
Django==4.0.6
django-allauth==0.51.0
django-autoslug==1.9.8
django-countries==7.3.2
django-crispy-forms==1.14.0
django-debug-toolbar==3.5.0
django-embed-video==1.4.4
djangorestframework==3.13.1
djlint==1.4.0
docopt==0.6.2
gunicorn==20.1.0
html-tag-names==0.1.2
html-void-elements==0.1.0
idna==3.3
importlib-metadata==4.12.0
isort==5.10.1
lazy-object-proxy==1.7.1
mccabe==0.7.0
mypy-extensions==0.4.3
oauthlib==3.2.0
packaging==21.3
pathspec==0.9.0
paypal-checkout-serversdk==1.0.1
paypalhttp==1.0.1
Pillow==9.2.0
pillow-heif==0.4.0
pip-upgrader==1.4.15
platformdirs==2.5.2
psycopg==3.0.15
psycopg-binary==3.0.15
pycodestyle==2.8.0
pycparser==2.21
PyJWT==2.4.0
pylint==2.14.4
pylint-celery==0.3
pylint-django==2.5.3
pylint-plugin-utils==0.7
pymemcache==3.5.2
pyOpenSSL==22.0.0
pyparsing==3.0.9
python-decouple==3.6
python-monkey-business==1.0.0
python3-openid==3.2.0
pytz==2022.1
PyYAML==6.0
regex==2022.6.2
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.2
terminaltables==3.1.10
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.0
tqdm==4.64.0
typing_extensions==4.3.0
urllib3==1.26.9
whitenoise==6.2.0
wrapt==1.14.1
zipp==3.8.0
